Updating scanner not to be a class. Moved source to src folder
